only destroy content value if its nil

+++ app/models/cms/content_attributes/concerns/reference_list.rb
@@ -20,7 +20,7 @@ module Cms
               if objs.present?
                 assign_value(serialize( objs.map(&:id) ))
               else
-                content_value.destroy
+                content_value.destroy if content_value
               end
             end
           end
